move many tool functions to tools.py
